// src/app/Models/base/mvtdetail.ts
export class mvtdetail {
    mv_mmc: number;
    jl_code: number;
    mv_folio: number;
    mv_lig: number;
    companyId: number;
    mv_montd: number;
    mv_cheq: string;
    mv_aac: number;
    mV_UTILIS: string;
    mv_compt: string;
    mV_jopt: number;
    mv_montc: number;
    mv_piece: number;
    mv_littrage: number;
    mv_date: Date;
    client: string;
    annee: number;
    fACTURE: number;
    mv_libop: string;
    mv_datec: Date;

}